Colour voting added to the rule list and experiment.py created

+++ no_ssh_versions/simulation.py
@@ -15,20 +15,79 @@ import string
 
 class Simulator:
 
-    def __init__(self,config_file, output):
+    def do_table_dicts(self):
+        #MAKE rule->id dict
+        rule_to_id = {}
+        for node in self.nodes:
+            rule = node[1]
+            instance = node[0]
+            if rule in rule_to_id:
+                rule_to_id[rule].append(instance)
+            else:
+                rule_to_id[rule] = [instance]
+
+        #MAKE rule->tables dict
+        rule_to_table_dic_list = {}
+        rule_to_tables = {}
+        for rule in rule_to_id:
+             dic_list = get_table_dict(rule,'persistent')
+             rule_to_table_dic_list[rule] = dic_list
+             table_names = []
+             for table in dic_list:
+                 table_names.append(table['name'])
+             rule_to_tables[rule] = table_names
+
+             
+        self.node_to_persistent = {}
+        for node in self.nodes:
+            rule = node[1]
+            instance = node[0]
+            self.node_to_persistent[instance] = rule_to_tables[rule]
+        
+        #List of tuples of the form ([Table dicts],[IDs])
+        self.template_list = []
+        for rule in rule_to_id:
+            table_dic_list = rule_to_table_dic_list[rule]
+            ids = rule_to_id[rule]
+            self.template_list.append((table_dic_list,ids))
+
+
+                    
+    def __init__(self,config_file, output, hardcoded_topology):
+        self.hardcoded_topology = hardcoded_topology
+
         #The file to which the output should be written
         self.output = output
 
         self.config = MyXMLParser(config_file)
-                
+        
+        self.one_rule = self.config.one_rule
         #List of all the instances to start
         self.nodes = self.config.nodes
         self.limit = self.config.limit
         self.rule = self.config.rule
-        
+
+        #self.do_table_dicts()
+ 
         #List of tables by type
-        self.tran_table_dict = get_table_dict(self.rule,'transport')
-        self.persistent_table_dict = get_table_dict(self.rule,'persistent')
+        if self.one_rule:
+            self.tran_table_dict = get_table_dict(self.rule,'transport')
+            self.pers_table_dict = get_table_dict(self.rule,'persistent')
+        else:
+            self.pers_table_dict = []
+            self.tran_table_dict = []
+            rules_seen = []
+                       
+            for node in self.nodes:
+                instance = node[0]
+                rule = node[1]
+                if rule not in rules_seen:
+                    for table in get_table_dict(rule,'transport'):
+                        self.tran_table_dict.append(table)
+                    for table in get_table_dict(rule,'persistent'):
+                        self.pers_table_dict.append(table)
+                rules_seen.append(rule)
+
 
         #List of (node1,node2) tuples representing
         #bidirectional links
@@ -65,22 +124,23 @@ class Simulator:
     def run_simulation(self):
         start_time = time.time()
         self.start_nodes()
-        start_time = time.time() - start_time
+        node_start_time = time.time() - start_time
         print 'Starting network with {0} nodes took: {1} secods'\
-              .format(len(self.nodes),start_time)
+              .format(len(self.nodes),node_start_time)
   
         watcher_monitor = WatcherMonitor(self.limit)
         #Launch a thread which watches the output of the engines
-        watcher_thread = Watcher(self.process_dict,self.config,watcher_monitor,self.persistent_table_dict)
+        watcher_thread = Watcher(self.process_dict,self.config,watcher_monitor,self.pers_table_dict,self.tran_table_dict)
         watcher_thread.start()
         
         #Do pre inputs in parallel
         self.input_pre()
        
         #Set the topology
-        self.set_topology()
+        if not self.hardcoded_topology:
+            self.set_topology()
 
-        self.start_evaluations()
+            self.start_evaluations()
 
         #Do post inputs
         self.input_post()
@@ -92,11 +152,21 @@ class Simulator:
 
         if watcher_monitor.hit_limit():
             print "HIT LIMIT"
+            self.hit_limit = True
         else:
             print "CONVERGENCE REACHED"
-               
+            self.hit_limit = False
+
+        self.evaluations = watcher_monitor.evaluations
         self.node_states = watcher_thread.stop()
+        lamport_start = time.time()
         self.lost_list, self.nr_received, self.nr_sent = lamport_transformation(self.node_states)
+        lamport_duration = time.time() - lamport_start
+        print "Lamport took {0} seconds".format(lamport_duration)
+        try:
+            percentage_lost = float((self.nr_sent - self.nr_received))/self.nr_sent
+        except ZeroDivisionError:
+            percentage_lost = 0
 
         f = open('output','w')
         for node in self.node_states:
@@ -106,11 +176,30 @@ class Simulator:
                 f.write(state.__str__())
         f.close()
 
+        write_to_file_start = time.time()
+        self.simulation_time = time.time() - start_time
         self.make_xml()
         f = open(self.output,'w')
         f.write(self.result)
         f.close()
+        write_duration = time.time() - write_to_file_start
+        print "Write to file: {0}s".format(write_duration)
 
+        size_of_file = os.path.getsize(self.output)
+
+        time_taken = time.time() - start_time
+        print "Took {0} seconds to run the whole simulation".format(time_taken)
+        print start_time
+        print time.time()
+        
+
+        time_taken = int(time_taken)
+        percentage_lost = round(percentage_lost * 100,2)
+        f = open('experiment_data', 'a')
+        to_write = '{0} & {1} & {2} & {3}\n'.format(self.evaluations, time_taken, self.nr_sent, percentage_lost)
+#to_write = 'STATE TRANSITIONS: {0},TOTAL TIME: {6}, SENT MESSAGES: {1}, PERCENTAGE LOST: {2}, LAMPORT TIME: {3}, WRITE TIME: {4}, WRITE SIZE: {5}\n'.format(self.evaluations, self.nr_sent, percentage_lost, lamport_duration, write_duration, size_of_file, time_taken)
+        f.write(to_write)
+        f.close()
 
     #==== HELPER FUNCTIONS FOR SIMULATION ====
     #=========================================
@@ -119,11 +208,10 @@ class Simulator:
         for ins in self.nodes:
             instance = ins[0]
             rule = ins[1]
+            
             cmd = 'dsmengine -namespace andrei -instance {0} {1}'\
-                   .format(instance,rule)
-                   
+                   .format(instance,rule)       
 
-            #print "Calling command {0}".format(cmd)
             self.process_dict[instance] = subprocess.Popen(cmd, shell=True,\
                                               stdout=subprocess.PIPE,\
                                               stderr=subprocess.PIPE,\
@@ -140,8 +228,9 @@ class Simulator:
             for ins in self.nodes:
                 instance = ins[0]
                 try:
+                    
                     l = self.process_dict[instance].stderr.readline()
-
+                    
                     if 'DSMEngine engine started'.format(instance) in l:
                         self.started += 1
                         
@@ -159,52 +248,50 @@ class Simulator:
             cmd = 'tuple insert andrei {0} start dummy=1'.format(instance)
             ins = Inserter(cmd)
             ins.start()
+            ins.join()
+            inserters.append(ins)
+        #for inserter in inserters:
+        #    inserter.join()
             
-        for inserter in inserters:
-            inserter.join()
         print "DONE: Inserting the start tuples"
 
     def input_pre(self):
         print "DOING: inputs before topology"
         inserters = []
         for inp in self.pre_inputs:
-            #Create the csv file
-            f = open('csv{0}'.format(inp['instance']),'w')
-            f.write(inp['var_names'] + '\n')
-            for row in inp['rows']:
-                f.write(row + '\n')
-            f.close()
-            
-            cmd = 'tuple insert andrei {0} {1} csv{2}'\
-                  .format(inp['instance'],inp['table_name'],inp['instance'])
 
+            if inp['type'] == 'csv':
+            #Create the csv file
+                f = open('csv{0}'.format(inp['instance']),'w')
+                f.write(inp['var_names'] + '\n')
+                for row in inp['rows']:
+                    f.write(row + '\n')
+                f.close()
+            
+                cmd = 'tuple insert andrei {0} {1} csv{2}'\
+                      .format(inp['instance'],inp['table_name'],inp['instance'])
+
+
+            elif inp['type'] == 'normal':
+                cmd ='tuple insert andrei {0} {1} {2}={3}'\
+                     .format(inp['instance'],inp['table_name'],
+                             inp['var_name'],inp['value'])
+
+            #Feed command to inserters
             ins = Inserter(cmd)
             ins.start()
             inserters.append(ins)
+
 
         for ins in inserters:
             ins.join()
 
         for inp in self.pre_inputs:
         #Delete the file
-            os.remove('csv{0}'.format(inp['instance']))
+            if inp['type'] == 'csv':
+                os.remove('csv{0}'.format(inp['instance']))
 
-        print "DONE: Inputs before topology"
-
-    #PARALLEL TOPOLOGY - NETWORK OVERLOADS -> MESSAGES LOST
-    def set_topology_parallel(self):
-        inserters = []
-        for link in self.links:
-            to_insert = [(link[0],link[1]),(link[1],link[0])]
-            for insert in to_insert:
-                print "Doing insert of {0}".format(insert)
-                inserter = LinkInserter(insert)
-                inserters.append(inserter)
-                inserter.start()
-        
-        for inserter in inserters:
-            inserter.join()
-            
+        print "DONE: Inputs before topology"    
     
     def input_post(self):
         print "DOING: Post Inputs"
@@ -312,11 +399,13 @@ class Simulator:
             else:
                 dic[link[0]] = [link[1]]
 
+            '''
             if link[1] in dic:
                 dic[link[1]].append(link[0])
             else:
                 dic[link[1]] = [link[0]]
-        
+                '''
+
         inserters = []
         files = []
         #create a CSV file and give it to an inserter
@@ -363,7 +452,7 @@ class Simulator:
                         if 'Tuples sent' in line:
                            print "LINK INSERTED: {0}".format(insert)
                            success = True
-                           continue
+                           continuel
                         if 'Tuples NOT sent' in line:
                            print "LINK FAILED: {0}".format(insert)
                            continue
@@ -373,7 +462,7 @@ class Simulator:
         result = Element('result')
         #Get info about the declarations at the start
         messages_elem = do_declarations(self.tran_table_dict,'messages')
-        tables_elem = do_declarations(self.persistent_table_dict,'tables')
+        tables_elem = do_declarations(self.pers_table_dict,'tables')
 
         result.append(messages_elem)
         result.append(tables_elem)
@@ -384,6 +473,15 @@ class Simulator:
         append_messages_lost(messages_lost_elem,self.lost_list)
         result.append(messages_lost_elem)
         
+        outcome_elem = Element('outcome')
+        outcome_elem.attrib['transitions'] = str(self.evaluations)
+        outcome_elem.attrib['time'] = str(int(self.simulation_time))
+        if self.hit_limit:
+            outcome_elem.attrib['outcome'] = 'Hit Limit'
+        else:
+            outcome_elem.attrib['outcome'] = 'Converged'
+        result.append(outcome_elem)
+
         states_elem = do_states(self.node_states)
         result.append(states_elem)
 
@@ -398,6 +496,9 @@ if __name__ == "__main__":
     parser.add_option("-c", "--config", dest="config",\
                      help="xml file to load configuration")
     '''                 
+    parser.add_option("-a", "--hardcoded",
+                      action="store_true", dest="hardcoded_topology", default=False,
+                      help="The topology is hardcoded already")
 
     (options,args) = parser.parse_args()
     if len(args) != 2:
@@ -411,8 +512,9 @@ if __name__ == "__main__":
     if output[0] == '"':
         output = output[1,len(output-1)]
 
+    hardcoded_topology = options.hardcoded_topology
     time.sleep(5)
-    simulator = Simulator(config,output)
+    simulator = Simulator(config,output,hardcoded_topology)
     
     simulator.run_simulation()
     time.sleep(5)
